BSP-154: Removing classes that were moved into BSP commons lib
