+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/bulkscan/transformations/BulkScanFormTransformer.java
@@ -1,8 +1,8 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.transformations;
 
 import uk.gov.hmcts.reform.bsp.common.error.UnsupportedFormTypeException;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.transformation.in.ExceptionRecord;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.in.OcrDataField;
+import uk.gov.hmcts.reform.bsp.common.model.transformation.in.ExceptionRecord;
+import uk.gov.hmcts.reform.bsp.common.model.validation.in.OcrDataField;
 
 import java.util.List;
 import java.util.Map;
