+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/bulk/scan/validation/FormAValidatorTest.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.service.bulk.scan.validatio
 
 import org.junit.Before;
 import org.junit.Test;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.in.OcrDataField;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.out.OcrValidationResult;
+import uk.gov.hmcts.reform.bsp.common.model.validation.in.OcrDataField;
+import uk.gov.hmcts.reform.bsp.common.model.validation.out.OcrValidationResult;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.bulkscan.validation.FormAValidator;
 
 import java.util.ArrayList;
@@ -14,14 +14,13 @@ import static java.util.Collections.emptyList;
 import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.out.ValidationStatus.SUCCESS;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.out.ValidationStatus.WARNINGS;
+import static uk.gov.hmcts.reform.bsp.common.model.validation.out.ValidationStatus.SUCCESS;
+import static uk.gov.hmcts.reform.bsp.common.model.validation.out.ValidationStatus.WARNINGS;
 
 public class FormAValidatorTest {
 
     private final FormAValidator classUnderTest = new FormAValidator();
     private List<OcrDataField> listOfAllMandatoryFields;
-    private OcrDataField validD8paymentMethod = new OcrDataField("D8PaymentMethod", "Cheque");
 
     @Before
     public void setup() {
